Fix UB in `LinkedList::len`
According to the [Rust Reference](https://doc.rust-lang.org/stable/reference/behavior-considered-undefined.html), dereferencing a null pointer is considered UB. When UB is encountered, anything is allowed to happen. Therefore, UB must be avoided.

This function is not unsafe, and must therefore never exhibit UB for any input provided to it.

This commit fixes the UB by always returning 0.