Optimize LinkedList::len
Local benchmarks show that when using this list in a real-world application (web3 related), a significant amount of time is spent in `LinkedList::len`. After some extra profiling with `perf` it shows that most time is spent in useless instructions after the `unreachable!`. Since these are not reachable, they can simply be deleted.

```
       │     .map_or(0, |node| unsafe { node.as_ref().next.len() + 1 }):
 87.50 │       movzbl  (%r12),%edx
  2.08 │       test    %dl,%dl
       │     → je      Node::as_ref
       │       mov     %r12,%rcx
```